+++ client/package.json
@@ -10,15 +10,18 @@
     "@types/jest": "^27.0.2",
     "@types/node": "^16.11.4",
     "@types/react": "^17.0.31",
+    "@types/react-datepicker": "^4.1.7",
     "@types/react-dom": "^17.0.10",
     "@types/react-redux": "^7.1.20",
     "@types/redux": "^3.6.0",
     "@types/styled-components": "^5.1.15",
     "axios": "^0.23.0",
     "classnames": "^2.3.1",
+    "date-fns": "^2.25.0",
     "eslint": "^7.32.0",
     "prop-types": "^15.7.2",
     "react": "^17.0.2",
+    "react-datepicker": "^4.3.0",
     "react-dom": "^17.0.2",
     "react-redux": "^7.2.5",
     "react-router-dom": "^5.3.0",
