+++ client/src/components/Modal/Modal.tsx
@@ -7,6 +7,7 @@ const Modal = () => {
   const { loginModal, signupModal } = useSelector(
     (state: RootState) => state.modal
   );
+  console.log(loginModal, signupModal);
 
   return (
     <>
