+++ client/src/store/login-slice.ts
@@ -6,7 +6,6 @@ export interface authentification {
 
 const initialState: authentification = { isLogin: false };
 
-
 const authSlice = createSlice({
   name: 'authentification',
   initialState: initialState,
